Separado arquivos dos repositórios
