+++ src/utils/repositories/news.ts
@@ -1,0 +1,20 @@
+export const news = [
+  {
+    name: 'Engeplus',
+    url: 'https://www.engeplus.com.br/noticia/espaco-empreendedor/2022/aplicativo-promete-facilitar-happy-hour-em-criciuma-e-regiao',
+    title: 'Aplicativo promete facilitar Happy Hour em Criciúma e Região',
+    imgFilePath: '/images/que-barbada-engeplus.jpg',
+  },
+  {
+    name: 'CanalSC',
+    url: 'https://canalsc.com.br/aplicativo-chega-para-os-descomplicar-o-role/',
+    title: 'Aplicativo criado em Jaraguá do Sul chega para os descomplicar o rolê',
+    imgFilePath: '/images/que-barbada-canalsc.jpeg',
+  },
+  {
+    name: 'Noticenter',
+    url: 'https://www.noticenter.com.br/n.php?ID=31809&T=aplicativo-que-barbada-criado-em-jaragua-do-sul-chega-para-descomplicar-o-role',
+    title: "Aplicativo Que Barbada, criado em Jaraguá do Sul, chega para 'descomplicar o rolê'",
+    imgFilePath: '/images/que-barbada-noticenter.jpeg',
+  },
+]
