+++ src/utils/repositories/rawCustomers.ts
@@ -562,90 +562,3 @@ export const rawCustomers: Customer[] = [
       'https://quebarbada.s3.sa-east-1.amazonaws.com/logo/20221021084931094-posto-beer.jpeg',
   },
 ]
-
-export const testimonials = [
-  {
-    testimonial:
-      '"Praticamente toda semana pessoas comentam que conheceram o Meu Boteco ao baixar o Que Barbada no hotel"',
-    owner: 'Pedro - proprietário Meu Boteco',
-    imageFileName: 'customer_pedro_meuboteco.jpeg',
-    width: 390,
-    height: 532,
-  },
-  {
-    testimonial:
-      '"Promovemos um show no Sinners Pub e mais de 40 ingressos foram vendidos através do app Que Barbada"',
-    owner: 'Baco - proprietário Sinners Pub',
-    imageFileName: 'customer_baco_sinners.jpeg',
-    width: 390,
-    height: 474,
-  },
-  {
-    testimonial: '"No dia dos namorados tivemos 3 reservas efetuadas pelo app Que Barbada"',
-    owner: 'Magnos - proprietário Samambaia',
-    imageFileName: 'customer_magnos_samambaia.jpeg',
-    width: 354,
-    height: 409,
-  },
-  {
-    testimonial:
-      '"Quando as pessoas veem o adesivo do Que Barbada no monitor do caixa, comentam que vieram conhecer o meu bar porque viram no app"',
-    owner: 'Moacyr - proprietário Barbatum Brew Pub',
-    imageFileName: 'customer_moacyr_barbatum.jpeg',
-    width: 390,
-    height: 462,
-  },
-  {
-    testimonial:
-      '"Aqui no Sun7 River os clientes comentam que acompanham a nossa agenda de eventos e promoções pelo app Que Barbada"',
-    owner: 'Fernanda - proprietária Sun7 River Gastrobar',
-    imageFileName: 'customer_fernanda_sun7.jpg',
-    width: 320,
-    height: 320,
-  },
-]
-
-export const supporters = [
-  {
-    name: 'Programa NaSCer',
-    imageFileName: 'nascer.png',
-    url: 'https://fapesc.sc.gov.br/programa-nascer/',
-    width: 962,
-    height: 679,
-  },
-  {
-    name: 'TXM Methods',
-    imageFileName: 'txm.png',
-    url: 'https://www.txm-methods.com/',
-    width: 527,
-    height: 254,
-  },
-  {
-    name: 'Fapesc',
-    imageFileName: 'fapesc.png',
-    url: 'https://fapesc.sc.gov.br/',
-    width: 795,
-    height: 850,
-  },
-]
-
-export const news = [
-  {
-    name: 'Engeplus',
-    url: 'https://www.engeplus.com.br/noticia/espaco-empreendedor/2022/aplicativo-promete-facilitar-happy-hour-em-criciuma-e-regiao',
-    title: 'Aplicativo promete facilitar Happy Hour em Criciúma e Região',
-    imgFilePath: '/images/que-barbada-engeplus.jpg',
-  },
-  {
-    name: 'CanalSC',
-    url: 'https://canalsc.com.br/aplicativo-chega-para-os-descomplicar-o-role/',
-    title: 'Aplicativo criado em Jaraguá do Sul chega para os descomplicar o rolê',
-    imgFilePath: '/images/que-barbada-canalsc.jpeg',
-  },
-  {
-    name: 'Noticenter',
-    url: 'https://www.noticenter.com.br/n.php?ID=31809&T=aplicativo-que-barbada-criado-em-jaragua-do-sul-chega-para-descomplicar-o-role',
-    title: "Aplicativo Que Barbada, criado em Jaraguá do Sul, chega para 'descomplicar o rolê'",
-    imgFilePath: '/images/que-barbada-noticenter.jpeg',
-  },
-]
