+++ src/utils/repositories/supporters.ts
@@ -1,0 +1,23 @@
+export const supporters = [
+  {
+    name: 'Programa NaSCer',
+    imageFileName: 'nascer.png',
+    url: 'https://fapesc.sc.gov.br/programa-nascer/',
+    width: 962,
+    height: 679,
+  },
+  {
+    name: 'TXM Methods',
+    imageFileName: 'txm.png',
+    url: 'https://www.txm-methods.com/',
+    width: 527,
+    height: 254,
+  },
+  {
+    name: 'Fapesc',
+    imageFileName: 'fapesc.png',
+    url: 'https://fapesc.sc.gov.br/',
+    width: 795,
+    height: 850,
+  },
+]
