+++ src/utils/repositories/testimonials.ts
@@ -1,0 +1,41 @@
+export const testimonials = [
+  {
+    testimonial:
+      '"Praticamente toda semana pessoas comentam que conheceram o Meu Boteco ao baixar o Que Barbada no hotel"',
+    owner: 'Pedro - proprietário Meu Boteco',
+    imageFileName: 'customer_pedro_meuboteco.jpeg',
+    width: 390,
+    height: 532,
+  },
+  {
+    testimonial:
+      '"Promovemos um show no Sinners Pub e mais de 40 ingressos foram vendidos através do app Que Barbada"',
+    owner: 'Baco - proprietário Sinners Pub',
+    imageFileName: 'customer_baco_sinners.jpeg',
+    width: 390,
+    height: 474,
+  },
+  {
+    testimonial: '"No dia dos namorados tivemos 3 reservas efetuadas pelo app Que Barbada"',
+    owner: 'Magnos - proprietário Samambaia',
+    imageFileName: 'customer_magnos_samambaia.jpeg',
+    width: 354,
+    height: 409,
+  },
+  {
+    testimonial:
+      '"Quando as pessoas veem o adesivo do Que Barbada no monitor do caixa, comentam que vieram conhecer o meu bar porque viram no app"',
+    owner: 'Moacyr - proprietário Barbatum Brew Pub',
+    imageFileName: 'customer_moacyr_barbatum.jpeg',
+    width: 390,
+    height: 462,
+  },
+  {
+    testimonial:
+      '"Aqui no Sun7 River os clientes comentam que acompanham a nossa agenda de eventos e promoções pelo app Que Barbada"',
+    owner: 'Fernanda - proprietária Sun7 River Gastrobar',
+    imageFileName: 'customer_fernanda_sun7.jpg',
+    width: 320,
+    height: 320,
+  },
+]
